+++ README.md
@@ -6,11 +6,11 @@ SPDX-License-Identifier: Apache-2.0
 # Tanzu GemFire Examples
 
 This repo contains examples of apps that demonstrate the use of VMware Tanzu GemFire and can be used with the 
-[Tanzu GemFire Developer Guide website](https://tanzugemfire.dev).
+[Tanzu Developer Center](https://tanzu.vmware.com/developer/data/tanzu-gemfire/).
 Each example app resides in its own directory within this repository:
 
 - Hello, World! in `hello-world`
-- Example of look-aside caching pattern in `look-aside-cache`
+- Example of Cache-aside pattern in `cache-aside`
 - Example of implementing an HTTP session state cache in `session-state`
 - Example of converting a Redis Publish/Subscribe app to Tanzu GemFire in `convert-redis-to-gemfire/pub-sub`
 - Examples of converting Spring Session Redis apps to Tanzu GemFire in `convert-redis-to-gemfire/session-state`
