+++ cache-aside/manifest.yml
@@ -1,0 +1,11 @@
+#
+# Copyright 2019 - 2021 VMware, Inc.
+# SPDX-License-Identifier: Apache-2.0
+#
+
+---
+applications:
+  - name: bike-incidents-app
+    path: ./build/libs/cache-aside-0.0.1-SNAPSHOT.jar
+    services:
+      - <your-tanzu-gemfire-service>
