+++ cache-aside/src/main/java/com/vmware/tanzu/gemfire/cacheaside/config/CacheAsideApplicationConfig.java
@@ -1,0 +1,22 @@
+/*
+Copyright 2019 - 2021 VMware, Inc.
+SPDX-License-Identifier: Apache-2.0
+*/
+
+package com.vmware.tanzu.gemfire.cacheaside.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.gemfire.config.annotation.EnableCachingDefinedRegions;
+import org.springframework.geode.config.annotation.EnableClusterAware;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+@EnableCachingDefinedRegions
+@EnableClusterAware
+public class CacheAsideApplicationConfig {
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+}
