+++ cache-aside/src/main/java/com/vmware/tanzu/gemfire/cacheaside/domain/BikeIncident.java
@@ -1,0 +1,35 @@
+/*
+Copyright 2019 - 2021 VMware, Inc.
+SPDX-License-Identifier: Apache-2.0
+*/
+
+
+package com.vmware.tanzu.gemfire.cacheaside.domain;
+
+public class BikeIncident {
+    private String type;
+    private String title;
+    private String description;
+    private String address;
+
+    public String getType() {
+        return type;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+}
