+++ cache-aside/src/main/resources/application.properties
@@ -1,0 +1,12 @@
+#
+# Copyright 2019 - 2021 VMware, Inc.
+# SPDX-License-Identifier: Apache-2.0
+#
+
+spring.template.cache=false
+server.error.whitelabel.enabled=false
+bikewise.api.url=https://bikewise.org:443/api/v2/incidents?page=1&proximity_square=50&proximity=
+
+#Properties used to connect to Tanzu GemFire for Kubernetes
+#spring.data.gemfire.pool.locators: cache-aside-gemfire-cluster-locator-0.cache-aside-gemfire-cluster-locator.tanzu-gemfire[10334]
+#spring.data.gemfire.management.http.host: cache-aside-gemfire-cluster-locator-0.cache-aside-gemfire-cluster-locator.tanzu-gemfire
