--- look-aside-cache/build.gradle
@@ -1,48 +1,0 @@
-/*
- * Copyright 2019 - 2021 VMware, Inc.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-plugins {
-    id 'org.springframework.boot' version '2.4.1'
-    id 'io.spring.dependency-management' version '1.0.10.RELEASE'
-    id 'java'
-}
-
-group = 'com.vmware.tanzu.gemfire.look-aside-cache'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '1.8'
-
-repositories {
-    mavenCentral()
-}
-
-bootBuildImage{
-    imageName = "docker.io/[YOUR DOCKER USERNAME]/bike-incidents:${project.version}"
-}
-
-ext {
-    set('springGeodeVersion', "1.4.0")
-}
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.geode:spring-geode-starter'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'org.springframework.geode:spring-geode-starter-test'
-
-}
-
-dependencyManagement {
-    imports {
-        mavenBom "org.springframework.geode:spring-geode-bom:${springGeodeVersion}"
-    }
-}
-
-test {
-    useJUnitPlatform()
-}
-
-
-
