--- look-aside-cache/manifest.yml
@@ -1,11 +1,0 @@
-#
-# Copyright 2019 - 2021 VMware, Inc.
-# SPDX-License-Identifier: Apache-2.0
-#
-
----
-applications:
-  - name: bike-incidents-app
-    path: ./build/libs/look-aside-cache-0.0.1-SNAPSHOT.jar
-    services:
-      - <your-tanzu-gemfire-service>
