--- look-aside-cache/src/main/java/com/vmware/tanzu/gemfire/lookasidecache/config/LookAsideCacheApplicationConfig.java
@@ -1,22 +1,0 @@
-/*
-Copyright 2019 - 2021 VMware, Inc.
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package com.vmware.tanzu.gemfire.lookasidecache.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.gemfire.config.annotation.EnableCachingDefinedRegions;
-import org.springframework.geode.config.annotation.EnableClusterAware;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-@EnableCachingDefinedRegions
-@EnableClusterAware
-public class LookAsideCacheApplicationConfig {
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-}
